backend/pyproject.toml
[project]
name = "yc-hackathon-backend"
version = "0.1.0"
description = "Consolidated backend dependencies for MCP agents"
requires-python = ">=3.10"
dependencies = [
  "fastmcp>=2.11.0",
  "modelcontextprotocol>=0.1.0",
  "amadeus==8.0.0",
  "pydantic>=2.9.2",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.2",
  "requests",
  "openai>=1.40.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
  "flight_mcp_agent",
  "hotel_mcp_agent",
]
